start component refactor. move messages related stuff to messages folder
