+++ app/containers/App.jsx
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import Navigation from 'containers/Navigation';
-import Message from 'containers/Message';
+import Message from 'messages';
 import classNames from 'classnames/bind';
 import styles from 'css/main';
