+++ app/messages/index.js
@@ -1,0 +1,1 @@
+export Message from './container';
